refactor(gotestrunner): convert tests to suite
