+++ internal/gotestrunner/export_test.go
@@ -1,0 +1,8 @@
+package gotestrunner
+
+type ExportOverlayJSON = overlayJSON
+
+var ExportBuildSuiteCmd = buildSuiteCmd
+var ExportReadTeardownBudget = readTeardownBudget
+var ExportSplitTopLevelOr = splitTopLevelOr
+var ExportSuiteRunFilter = suiteRunFilter
